Added ability to delete user, updated README
This commit adds the functionality to delete a user
and a left network from the VPN concentrator. A user
or network may be deleted through stack updates, then
prefixing "delete:" to a command delmited list of entries
in it's appropriate parameter field.
For example, in the Users field:

delete: testuser,anotheruser

This would delete the user 'testuser' and 'anotheruser'
from the VPN concentrator. The same goes for networks.
See README for more details.